port: RE4_FORMAT_PRINTF audit for Ptr32<T>/BE<T> dropped through C varargs (bytes unchanged)
-Wformat on the host build (CMakeLists.txt) plus RE4_FORMAT_PRINTF (include/port/format_attr.h,
TARGET_PC-only) on eprintf/eprintf2/OSReport/OSPanic/sprintf/cLog::mes-err-warn found 5 more sites
beyond title.cpp's own fix (f465a298): room_jmp.cpp's roomJumpMove (info->name/person/person2,
Ptr32<char>) and one BE<u32> each in eff_sys.cpp (tpl->numDescriptors) and trans.cpp (d->nTex).
Cast at each call site, same pattern as title.cpp's existing fix; no-op on the original target
(fields are already char*/u32 there).

+++ src/game/eff_sys.cpp
@@ -612,7 +612,7 @@ int espTexRegist(TEXPalette* tpl, EspAnmData* anm, u8 id, u32 owner)
         return 0;
     }
     if (anm->Frames != tpl->numDescriptors) {
-        pLog->err(0, 0, "ESP : ID[%02x] TEX/ANM ptn num diff[%d / %d]", id, tpl->numDescriptors, anm->Frames);
+        pLog->err(0, 0, "ESP : ID[%02x] TEX/ANM ptn num diff[%d / %d]", id, (u32) tpl->numDescriptors, anm->Frames); // (u32) cast: BE<u32> under TARGET_PC, varargs drop its conversion operator; no-op on the original target where the field is plain u32 (bytes unchanged)
         return 0;
     }
     EspCalcTplAddr(tpl);
